requirements.txt
# FastMCP and FastAPI dependencies
fastmcp>=2.3.0
uvicorn[standard]>=0.30.0

# Google Cloud SDK libraries
google-cloud-compute>=1.20.0
google-cloud-run>=0.10.0
google-cloud-resource-manager>=1.12.0

# Configuration and environment management
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# Development dependencies
black>=24.0.0
ruff>=0.5.0
mypy>=1.11.0

# Google Auth
google-auth>=2.30.0
